requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aioice==0.10.1
aiortc==1.11.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asyncer==0.0.8
attrs==25.3.0
autogen==0.8.7
autogen-agentchat==0.5.3
autogen-core==0.5.3
autogen-ext==0.5.3
av==14.3.0
azure-core==1.33.0
azure-identity==1.21.0
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudevents==1.11.0
colorama==0.4.6
cryptography==44.0.2
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
easyocr==1.7.2
fastapi==0.115.12
filelock==3.18.0
flake8==7.2.0
frozenlist==1.6.0
fsspec==2025.3.2
google-crc32c==1.7.1
griffe==1.7.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
ifaddr==0.2.0
imageio==2.37.0
importlib_metadata==8.6.1
iniconfig==2.1.0
isodate==0.7.2
isort==6.0.1
Jinja2==3.1.6
jiter==0.9.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-path==0.3.4
jsonschema-specifications==2024.10.1
lazy-object-proxy==1.11.0
lazy_loader==0.4
MarkupSafe==3.0.2
mccabe==0.7.0
mcp==1.6.0
mercadopago==2.3.0
more-itertools==10.6.0
mpmath==1.3.0
msal==1.32.0
msal-extensions==1.3.1
multidict==6.4.3
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
numpy==1.26.4
openai==1.75.0
openai-agents==0.0.11
openapi-core==0.19.5
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opentelemetry-api==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
packaging==24.2
pandas==2.2.3
parse==1.20.2
pathable==0.4.4
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
prance==23.6.21.0
propcache==0.3.1
protobuf==5.29.4
pyautogen==0.8.7
pybars4==0.9.13
pyclipper==1.3.0.post6
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pyee==13.0.0
pyflakes==3.3.2
PyJWT==2.10.1
pylibsrtp==0.12.0
PyMeta3==0.5.1
pyOpenSSL==25.0.0
pytest==8.3.5
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rpds-py==0.24.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-image==0.25.2
scipy==1.15.2
semantic-kernel==1.28.1
shapely==2.1.0
six==1.17.0
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.46.2
sympy==1.13.3
termcolor==3.0.1
tifffile==2025.3.30
tiktoken==0.9.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.1
websockets==15.0.1
Werkzeug==3.1.1
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
